% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constAvBetaTSS.R
\name{constAvBetaTSS}
\alias{constAvBetaTSS}
\title{Construct a DNAm matrix at gene-level}
\usage{
constAvBetaTSS(beta.m, type = c("450k", "850k"))
}
\arguments{
\item{type}{This specifies whether data is 450k or 850k.}
\item{beta.m}{A DNAm beta-valued data matrix with columns labeling samples
  and rows labeling CpGs. At present we only support Illumina 450k or
  850k methylation beadarrays.}
}
\value{
The DNAm data matrix at the gene-level.
}
\description{
This function takes as input a DNAm data matrix generated with
Illumina 450k or 850k technology and outputs a DNAm data matrix
at the gene level, by computing the average DNAm over a window
200bp upstream of TSS, or if not available over 1st exon probes.
}
\examples{
data(testDNAm);
avDNAm.m <- constAvBetaTSS(testDNAm.m,type="450k");



}
\references{
Teschendorff AE, Zhu T, Breeze CE, Beck S.
\emph{Cell-type deconvolution of bulk tissue DNA methylomes 
from single-cell RNA-Seq data}
Genome Biol.2020

Jiao Y, Widschwendter M, Teschendorff AE. 
\emph{A systems-level integrative framework for genome-wide DNA methylation
and gene expression data identifies differential gene expression modules
under epigenetic control.}
Bioinformatics (2014) 30(16).
doi:\href{https://doi.org/10.1093/bioinformatics/btu316}{
10.1093/bioinformatics/btu316}.
}
