% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstExpRef.R
\name{ConstExpRef}
\alias{ConstExpRef}
\title{Construct mRNA expression reference matrix}
\usage{
ConstExpRef(exp.m, celltype.idx, namesCellT.v, markspecTH.v = NULL, ncores = 8)
}
\arguments{
\item{exp.m}{The normalized scRNA-Seq data matrix with rows labeling unique genes (either human
gene symbol or Entrez gene ID) and columns labeling cells. Missing values are not
allowed and values should be positive or zero. Importantly, the value corresponding
zero read counts should also be zero. Typically, this matrix will be obtained by
scaling the readcounts, adding a pseudocount of +1 and then taking the log2 transform.}

\item{celltype.idx}{An integer index vector with as many elements as there are columns in 'exp.m',
specifying the cell-type of each cell. Each unique integer in the vector defines
one cell-type and integers must be sequential, i.e. no gaps are allowed. For instance,
for 4 cell-types, the integers would be 1 to 4.}

\item{namesCellT.v}{A vector of unique names for the cell-types, of length equal to the number
of unique elements in celltype.idx. The order of the names must match the numerical
integers in celltype.idx.}

\item{markspecTH.v}{This is the marker specificity score (MSS) threshold for each cell-type, i.e.
for a given marker gene and cell-type, the minimum number of other cell-types where
we require this gene not to be expressed. For n cell-types, the maximum MSS value is
n-1 and minimum value should be 1. The parameter must be a vector with one threshold
value for each cell-type. If 'NULL' (the default option), each entry will be set to n-1.}

\item{ncores}{The number of processing cores to use.}

}
\value{
  A list with two elements.
  \item{ref}{a list itself also with 2 elements, giving the reference expression matrices defined using averages (\code{av}) or medians (\code{med}).}
  \item{markers}{a list of matrices, one for each cell-type listing for each marker gene in the reference, their Wilcoxon rank sum statistics, median value in each cell-type and their marker specificity score.}
}
\description{
This function takes as input a tissue-specific scRNA-Seq atlas and 
produces an expression reference matrix for the cell-types specified
by the user
}
\examples{
data(lungSS2mca1)
out.l <- ConstExpRef(lungSS2mca1.m,celltypeSS2.idx,celltypeSS2.v,markspecTH.v=rep(3,4));
print(head(out.l$ref$med));

}
\references{
Zhu T, Breeze CE, Beck S, Teschendorff AE
\emph{Cell-type deconvolution of bulk tissue DNA methylomes 
from single-cell RNA-Seq data}
Submitted.
}
