% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstMergedDNAmRef.R
\name{ConstMergedDNAmRef}
\alias{ConstMergedDNAmRef}
\title{Construct merged DNAm reference matrix}
\usage{
ConstMergedDNAmRef(ref1.m, ref2.m)
}
\arguments{
\item{ref1.m}{The first DNAm reference matrix, say generated from the SCM2 database}

\item{ref2.m}{The second DNAm reference matrix, say generated from the RMAP database}
}
\value{
The merged DNAm reference matrix
}
\description{
This function takes as input two separate imputed DNAm reference matrices
and merges them to arrive at a unique final DNAm reference matrix.
}
\examples{
data(lungSS2mca1)
out.l <- ConstExpRef(lungSS2mca1.m,celltypeSS2.idx,celltypeSS2.v,markspecTH.v=rep(3,4));
refDNAm1.m <- ImputeDNAmRef(out.l$ref$med,db="SCM2",geneID="SYMBOL");
refDNAm2.m <- ImputeDNAmRef(out.l$ref$med,db="RMAP",geneID="SYMBOL");
refDNAm.m <- ConstMergedDNAmRef(refDNAm1.m,refDNAm2.m);
print(head(refDNAm.m));



}

\references{
Teschendorff AE, Zhu T, Breeze CE, Beck S.
\emph{Cell-type deconvolution of bulk tissue DNA methylomes 
from single-cell RNA-Seq data}
Genome Biol.2020

Zhu T, Liu J, Beck S, Pan S, Capper D, Lechner M, Thirlwell C, Breeze
CE, Teschendorff AE.
\emph{A pan-tissue DNA methylation atlas enables in silico decomposition of human tissue methylomes at cell-type resolution}
Nat Methods 2022
}
