\docType{data}
\name{Pancreas9ctRef}
\alias{Pancreas9ctRef}
\title{Pancreas9ct mRNA expression and DNAm reference matrix}
\format{Two matrices representing the mRNA and DNAm reference matrices
}

\usage{
data(PancreasRef9ct)
}
\description{
The expression reference matrix is defined over 941 marker genes and 9
cell-types (acinar, alpha, beta, delta, ductal, endothelial, gamma, immune-cell,stellate). The DNAm reference matrix is defined over 192 marker gene promoters,and the same 9 cell-types, with the last column labeling the confidence weight.
}
\details{
Upon loading the data object, we find the following:
\itemize{
  \item \code{exprefPancreas9ct.m}: the expression reference matrix
  \item \code{mrefPancreas9ct.m}: the DNAm reference matrix
}
}
\references{
}
\keyword{datasets}
